+++ srcs/core/maths/ft_fmod.c
@@ -6,7 +6,7 @@
 /*   By: rel-qoqu <rel-qoqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/03 08:44:08 by rel-qoqu          #+#    #+#             */
-/*   Updated: 2025/06/03 08:56:02 by rel-qoqu         ###   ########.fr       */
+/*   Updated: 2025/07/20 14:07:16 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,14 +18,14 @@ double	ft_fmod(const double x, const double y)
 	double	n;
 
 	if (y == 0.0)
-		return (errno = EDOM, NAN);
+		return (errno = EDOM, (double)NAN);
 	if (ft_is_nan(x) || ft_is_nan(y))
-		return (NAN);
+		return ((double)NAN);
 	if (ft_is_inf(x))
 	{
 		if (ft_is_inf(y))
-			return (errno = EDOM, NAN);
-		return (errno = EDOM, NAN);
+			return (errno = EDOM, (double)NAN);
+		return (errno = EDOM, (double)NAN);
 	}
 	if (ft_is_inf(y))
 		return (x);
