+++ srcs/core/maths/ft_rad_to_deg.c
@@ -6,7 +6,7 @@
 /*   By: rel-qoqu <rel-qoqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/31 18:41:24 by rel-qoqu          #+#    #+#             */
-/*   Updated: 2025/05/31 23:02:11 by rel-qoqu         ###   ########.fr       */
+/*   Updated: 2025/07/20 14:07:06 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 double	ft_rad_to_deg(double radians)
 {
 	if (ft_is_nan(radians))
-		return (NAN);
+		return ((double)NAN);
 	else if (ft_is_inf(radians))
 	{
 		if (radians < 0.0)
-			return (-INFINITY);
-		return (INFINITY);
+			return (-(double)INFINITY);
+		return ((double)INFINITY);
 	}
 	return (radians * 180.0 / PI);
 }
